Tech: déplace la classe de base dans un module séparé
