+++ src/scripts/page/thematiques/choixTestDepistage.js
@@ -1,72 +1,10 @@
-import { hideElement, hideSelector, showElement } from '../../affichage'
-import {
-    getRadioValue,
-    toggleFormButtonOnRadioRequired,
-    uncheckAllRadio,
-} from '../../formutils'
+import { hideElement } from '../../affichage'
+import { getRadioValue, toggleFormButtonOnRadioRequired } from '../../formutils'
+import { Formulaire } from './formulaire'
 
 export function dynamiseLeChoixDuTest() {
     const formulaire = new FormulaireTestDepistage()
     formulaire.demarre()
-}
-
-class Formulaire {
-    constructor(prefixe) {
-        this.prefixe = prefixe
-    }
-
-    demarre(etape = 'demarrage') {
-        const form = document.querySelector(`#${this.prefixe}-${etape}-form`)
-        this.appelleGestionnaire(form, etape)
-    }
-
-    transitionneVersEtape(etape) {
-        const form = document.querySelector(`#${this.prefixe}-${etape}-form`)
-        showElement(form)
-        this.gereBoutonRetour(form)
-        this.appelleGestionnaire(form, etape)
-    }
-
-    appelleGestionnaire(form, etape) {
-        const gestionnaire = this.GESTIONNAIRES[etape]
-        gestionnaire(form)
-    }
-
-    gereBoutonRetour(form) {
-        const boutonRetour = form.querySelector('.back-button')
-        if (!boutonRetour) return
-        boutonRetour.addEventListener('click', (event) => {
-            event.preventDefault()
-            const etapePrecedente = boutonRetour.dataset.precedent
-            hideElement(form)
-            this.transitionneVersEtape(etapePrecedente)
-        })
-    }
-
-    afficheReponse(etape) {
-        const reponse = document.querySelector(`#${this.prefixe}-${etape}-reponse`)
-        showElement(reponse)
-        this.gereBoutonRefaire()
-    }
-
-    gereBoutonRefaire() {
-        const boutonRefaire = document.querySelector(`#${this.prefixe}-refaire`)
-        showElement(boutonRefaire)
-        boutonRefaire.addEventListener('click', (event) => {
-            event.preventDefault()
-            hideElement(boutonRefaire)
-            hideSelector(document, '.statut')
-            uncheckAllRadio(document)
-            Array.from(
-                document.querySelectorAll('[data-initial-value]'),
-                (inputWithInitial) => {
-                    inputWithInitial.value = inputWithInitial.dataset.initialValue
-                    inputWithInitial.removeAttribute('data-initial-value')
-                }
-            )
-            this.transitionneVersEtape('symptomes')
-        })
-    }
 }
 
 class FormulaireTestDepistage extends Formulaire {
